requirements.txt
anyio==3.6.2
autopep8==2.0.0
azure-core==1.26.2
azure-storage-blob==12.14.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.85.2
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
isodate==0.6.1
msal==1.20.0
msrest==0.7.1
mysqlclient==2.1.1
numpy==1.24.0
oauthlib==3.2.2
pandas==1.5.2
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.4
